Cast operand to larger destination size to avoid possible overflow issues.

+++ src/zm_camera.cpp
@@ -63,7 +63,7 @@ Camera::Camera(
 {
   linesize = width * colours;
   pixels = width * height;
-  imagesize = height * linesize;
+  imagesize = static_cast<unsigned long long>(height) * linesize;
 
   Debug(2, "New camera id: %d width: %d line size: %d height: %d colours: %d subpixelorder: %d capture: %d",
       monitor->Id(), width, linesize, height, colours, subpixelorder, capture);
